requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.14.0
faunadb==4.5.0
python-dotenv==1.0.0
sentry-sdk==1.22.2
